test(storage): reuse git worktree fixture
